backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
psycopg2-binary>=2.9
Pillow>=10.0
django-cors-headers>=4.0.0
requests>=2.31.0
django-filter>=23.0
